+++ resources/views/components/form-section.blade.php
@@ -8,19 +8,19 @@
 
     <div class="mt-5 md:mt-0 md:col-span-2">
         <form @if($submit) wire:submit="{{ $submit }}" @endif>
-            <div
-                class="px-4 py-5 bg-gray-800/10 backdrop-blur-sm border border-white/20 shadow-sm {{ isset($actions) ? 'sm:rounded-tl-md sm:rounded-tr-md' : 'sm:rounded-md' }}">
-                <div class="grid grid-cols-6 gap-6">
-                    {{ $form }}
+            <div class="bg-gray-800/10 backdrop-blur-sm border border-white/20 shadow-sm overflow-hidden {{ isset($actions) ? 'sm:rounded-md' : 'sm:rounded-md' }}">
+                <div class="px-4 py-5">
+                    <div class="grid grid-cols-6 gap-6">
+                        {{ $form }}
+                    </div>
                 </div>
-            </div>
 
-            @if (isset($actions))
-                <div
-                    class="flex items-center justify-end px-4 py-3 bg-gray-800/10 backdrop-blur-sm border-t border-white/20 text-end sm:px-6 shadow-sm sm:rounded-bl-md sm:rounded-br-md">
-                    {{ $actions }}
-                </div>
-            @endif
+                @if (isset($actions))
+                    <div class="flex items-center justify-end px-4 py-3 border-t border-white/20 text-end sm:px-6">
+                        {{ $actions }}
+                    </div>
+                @endif
+            </div>
         </form>
     </div>
 </div>